EntriesParser added
